backend/requirements.txt
fastapi
uvicorn
python-dotenv
boto3
requests
python-jose[cryptography]
